x_auto_post/requirements.txt
google-generativeai
tweepy
schedule
python-dotenv
pillow
requests
